Update tests related to template definition listing API
